% Generated by roxygen2 (4.0.0): do not edit by hand
\name{copy_env}
\alias{copy_env}
\title{Copy one environment into another recursively.}
\arguments{
\item{to}{environment. The new environment.}

\item{from}{environment. The old environment.}
}
\description{
Copy one environment into another recursively.
}

