% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagerunner.R
\name{is.stagerunner}
\alias{is.stageRunner}
\alias{is.stagerunner}
\title{Check whether an R object is a stageRunner object}
\usage{
is.stagerunner(obj)

is.stageRunner(obj)
}
\arguments{
\item{obj}{any object.}
}
\value{
\code{TRUE} if the object is of class
   \code{stageRunner}, \code{FALSE} otherwise.
}
\description{
Check whether an R object is a stageRunner object
}

