% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_and_lists.R
\name{special_and_lists}
\alias{special_and_lists}
\title{AND two lists together with some regards for nesting}
\usage{
special_and_lists(list1, list2)
}
\arguments{
\item{list1}{a list.}

\item{list2}{a list.}
}
\value{
the and'ed list.
}
\description{
The structure of the lists should be the same. That is,
as a tree, the two lists should be isomorphic. For example,
\code{special_and_lists(list(a = FALSE, b = list(b = TRUE, c = FALSE)),
                        list(a = FALSE, b = list(b = FALSE, c = TRUE)))}
yields
\code{list(a = FALSE, b = list(b = FALSE, c = TRUE))}
and
\code{special_and_lists(list(a = FALSE, b = list(b = TRUE, c = FALSE)),
                        list(a = list(b = FALSE, c = TRUE), b = FALSE))}
yields
\code{list(a = list(b = FALSE, c = TRUE), b = list(b = TRUE, c = FALSE))}
}
\details{
Note that lists get ANDed based on *order*, not on key names (as this could
be ambiguous), so make sure the two lists have the same comparable key orders.
For example, \code{special_and_lists(list(a = TRUE, b = FALSE), list(b = FALSE, a = TRUE))}
would mistakenly return \code{list(a = TRUE, b = TRUE)}.
}
\examples{
\dontrun{
  stopifnot(identical(
    special_and_lists(list(a = FALSE, b = list(b = TRUE, c = FALSE)),
                      list(a = FALSE, b = list(b = FALSE, c = TRUE))),
    list(a = FALSE, b = list(b = FALSE, c = TRUE))
 ))

  stopifnot(identical(
    special_and_lists(list(a = FALSE, b = list(b = TRUE, c = FALSE)),
                      list(a = list(b = FALSE, c = TRUE), b = FALSE)),
    list(a = list(b = FALSE, c = TRUE), b = list(b = TRUE, c = FALSE))
 ))
}
}
\seealso{
\code{\link{special_or_lists}}
}

