% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner-internal.R
\name{stageRunner_.before_env}
\alias{stageRunner_.before_env}
\title{Get an environment representing the context directly before executing a given stage.}
\usage{
stageRunner_.before_env(stage_index)
}
\arguments{
\item{stage_index}{integer. The substage for which to grab the before
environment.}
}
\value{
a fresh new environment representing what would have been in
  the context as of right before the execution of that substage.
}
\description{
Get an environment representing the context directly before executing a given stage.
}
\note{
If there is a lot of data in the remembered environment, this function
  may be computationally expensive as it has to create a new environment
  with a copy of all the relevant data.
}

