% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner-internal.R
\name{stageRunner_.mark_finished}
\alias{stageRunner_.mark_finished}
\title{Mark a given stage as being finished.}
\usage{
stageRunner_.mark_finished(stage_index)
}
\arguments{
\item{stage_index}{integer. The index of the substage in this stageRunner.}
}
\description{
Mark a given stage as being finished.
}

