% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner_R6.R
\name{stageRunner_.root}
\alias{stageRunner_.root}
\title{Determine the root of the stageRunner.}
\usage{
stageRunner_.root()
}
\value{
the root of the stageRunner
}
\description{
Determine the root of the stageRunner.
}

