% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagerunner-internal.R
\name{stageRunner_.set_parents}
\alias{stageRunner_.set_parents}
\title{Set all parents for this stageRunner, and recursively}
\usage{
stageRunner_.set_parents()
}
\description{
Set all parents for this stageRunner, and recursively
}
