% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagerunner-objectdiff.R
\name{stageRunner_.set_prefixes}
\alias{stageRunner_.set_prefixes}
\title{Set all prefixes for child stageRunners.}
\usage{
stageRunner_.set_prefixes(prefix = "")
}
\arguments{
\item{prefix}{character. The prefix to assign to this stageRunner.}
}
\description{
When a stageRunner is used in conjunction with an
\code{objectdiff::tracked_environment}, we need to remember
the full nested tree structure. This function sets up the
\code{prefix} member of each sub-stageRunner recursively to enable
correct remembering functionality.
}
