\name{stageRunner__.clear_cache}
\alias{stageRunner__.clear_cache}
\title{Clear all caches in this stageRunner, and recursively.}
\usage{
stageRunner__.clear_cache(init = FALSE)
}
\arguments{
  \item{init}{logical. Internal argument.}
}
\description{
Clear all caches in this stageRunner, and recursively.
}

