% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/stage_runner.r
\name{stageRunner__.root}
\alias{stageRunner__.root}
\title{Determine the root of the stageRunner.}
\usage{
stageRunner__.root()
}
\value{
the root of the stageRunner
}
\description{
Determine the root of the stageRunner.
}

