\name{stageRunner__.set_parents}
\alias{stageRunner__.set_parents}
\title{Set all parents for this stageRunner, and recursively}
\usage{
stageRunner__.set_parents(init = FALSE)
}
\arguments{
  \item{init}{logical. Internal argument.}
}
\description{
Set all parents for this stageRunner, and recursively
}

