% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/stage_runner.r
\name{stageRunner__append}
\alias{stageRunner__append}
\title{Append one stageRunner to the end of another.}
\usage{
stageRunner__append(other_runner, label = NULL)
}
\arguments{
\item{other_runner}{stageRunner. Another stageRunner to append to the current one.}

\item{label}{character. The label for the new stages (this will be the name of the
newly appended list element).}
}
\description{
Append one stageRunner to the end of another.
}

