% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stageRunner__coalesce}
\alias{stageRunner__coalesce}
\title{Coalescing a stageRunner object is taking another stageRunner object
with similar stage names and replacing the latter's cached environments
with the former's.}
\arguments{
\item{other_runner}{stageRunner. Another stageRunner from which to coalesce.}
}
\description{
Coalescing a stageRunner object is taking another stageRunner object
with similar stage names and replacing the latter's cached environments
with the former's.
}

