% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stageRunner__has_key}
\alias{stageRunner__has_key}
\title{Whether or not the stageRunner has a key matching this input.}
\usage{
stageRunner__has_key(key)
}
\arguments{
\item{key}{ANY. The potential key.}
}
\value{
\code{TRUE} or \code{FALSE} accordingly.
}
\description{
Whether or not the stageRunner has a key matching this input.
}

