% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner-coalesce.R
\name{stageRunner_coalesce}
\alias{stageRunner_coalesce}
\title{Coalescing a stageRunner object is taking another stageRunner object
with similar stage names and replacing the latter's cached environments
with the former's.}
\usage{
stageRunner_coalesce(other_runner)
}
\arguments{
\item{other_runner}{stageRunner. Another stageRunner from which to coalesce.}
}
\description{
Coalescing a stageRunner object is taking another stageRunner object
with similar stage names and replacing the latter's cached environments
with the former's.
}
\note{
coalescing is ill-defined for stageRunner with unnamed stages,
   since it is impossible to tell when a stage has changed.
}

