% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner-has_key.R
\name{stageRunner_has_key}
\alias{stageRunner_has_key}
\title{Whether or not the stageRunner has a key matching this input.}
\usage{
stageRunner_has_key(key)
}
\arguments{
\item{key}{ANY. The potential key.}
}
\value{
\code{TRUE} or \code{FALSE} accordingly.
}
\description{
Whether or not the stageRunner has a key matching this input.
}

