% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner-next_stage.R
\name{stageRunner_next_stage}
\alias{stageRunner_next_stage}
\title{For stageRunners with caching, find the next unexecuted stage.}
\usage{
stageRunner_next_stage()
}
\value{
a character stage key giving the next unexecuted stage.
  If all stages have been executed, this returns \code{FALSE}.
  If the stageRunner does not have caching enabled, this will
  always return the first stage key (`'1'`).
}
\description{
For stageRunners with caching, find the next unexecuted stage.
}

