% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner_R6.R
\name{stageRunner_show}
\alias{stageRunner_show}
\title{Generic for printing stageRunner objects.}
\usage{
stageRunner_show(indent = 0)
}
\arguments{
\item{indent}{integer. Internal parameter for keeping track of nested
indentation level.}
}
\description{
Generic for printing stageRunner objects.
}

