% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagerunner-stage_names.R
\name{stageRunner_stage_names}
\alias{stageRunner_stage_names}
\title{Retrieve a flattened list of canonical stage names for a stageRunner object}
\usage{
stageRunner_stage_names()
}
\value{
a list of canonical stage names.
}
\description{
For example, if we have stages
  \code{stages = list(a = list(b = 1, c = 2), d = 3, e = list(f = 4, g = 5))}
then this method would return
  \code{list('a/b', 'a/c', 'd', 'e/f', 'e/g')}
}
