% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagerunner-transform.R
\name{stageRunner_transform}
\alias{stageRunner_transform}
\title{Transform the callable's of the terminal nodes of a stageRunner.}
\usage{
stageRunner_transform(transformation)
}
\arguments{
\item{transformation}{function. The function which transforms one callable
into another.}
}
\description{
Every terminal node in a stageRunner is of type stageRunnerNode.
These each have a callable, and this method transforms those
callables in the way given by the first argument.
}
