% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner_R6.R
\docType{class}
\name{stageRunnerNode}
\alias{stageRunnerNode}
\alias{stageRunnerNode_}
\title{Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "stageRunnerNode"
 $ public    :List of 15
  ..$ callable    : NULL
  ..$ cached_env  : NULL
  ..$ .context    : NULL
  ..$ .parent     : NULL
  ..$ executed    : logi FALSE
  ..$ initialize  :function (.callable, .context = NULL)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 729 18 732 5 18 5 729 732
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
  ..$ run         :function (..., .cached_env = NULL, .callable = callable)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 733 11 747 5 11 5 733 747
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
  ..$ around      :function (other_node)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 750 14 778 5 14 5 750 778
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
  ..$ overlay     :function (other_node, label = NULL, flat = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 780 15 796 5 15 5 780 796
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
  ..$ transform   :function (transformation)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 797 17 800 5 17 5 797 800
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
  ..$ was_executed:function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 801 20 801 42 20 42 801 801
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
  ..$ parent      :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 802 16 802 50 16 50 802 802
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
  ..$ children    :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 803 16 803 32 16 32 803 803
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
  ..$ show        :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 804 16 804 87 16 87 804 804
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
  ..$ index       :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 807 16 811 5 16 5 807 811
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10579b808> 
 $ private   : NULL
 $ active    : NULL
 $ inherit   : NULL
 $ parent_env:<environment: namespace:stagerunner> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
stageRunnerNode_
}
\arguments{
\item{fn}{function. This will be wrapped in an environment.}

\item{parent_obj}{stageRunner. The enclosing stageRunner object.}

\item{parent_env}{environment. The parent environment of the created
\code{stageRunnerNode} object. The default is the calling
environment (i.e., \code{parent.frame()}).}
}
\value{
an environment with some additional attributes for
  navigating in a tree-like structure.
}
\description{
Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).
}
\keyword{datasets}

