% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stageRunnerNode-around.R
\name{stageRunnerNode_around}
\alias{stageRunnerNode_around}
\title{Wrap a stageRunnerNode callable with another callable.}
\usage{
stageRunnerNode_around(other_node)
}
\arguments{
\item{other_node}{stagerunner or stageRunnerNode.}
}
\value{
\code{TRUE} or \code{FALSE} according as the wrapping was
   successful.
}
\description{
Wrap a stageRunnerNode callable with another callable.
}
\examples{
\dontrun{
node1 <- stageRunnerNode(function(e) print(2))
node2 <- stageRunnerNode(function(e) { print(1); yield(); print(3); })
node1$around(node2)
node1$run() # Will print 1 2 3
# Notice the provided "yield" keyword, which allows calling the
# node that is being wrapped.
}
}

