% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stageRunnerNode-overlay.R
\name{stageRunnerNode_overlay}
\alias{stageRunnerNode_overlay}
\title{Append one stageRunnerNode around another.}
\usage{
stageRunnerNode_overlay(other_node, label = NULL, flat = FALSE)
}
\arguments{
\item{other_node}{stagerunner or stageRunnerNode.}

\item{label}{character. Under the hood, this will be the "stage name"
for the stage represented by the \code{other_node} in the
automatically generated new stageRunner used as this node's
callable (assuming \code{flat} is \code{FALSE}).}

\item{flat}{logical. If \code{TRUE}.}
}
\value{
\code{TRUE} or \code{FALSE} according as the wrapping was
   successful.
}
\description{
Append one stageRunnerNode around another.
}
\examples{
\dontrun{
node1 <- stageRunnerNode(function(e) print(2))
node2 <- stageRunnerNode(function(e) { print(1); yield(); print(3); })
node1$around(node2)
node1$run() # Will print 1 2 3
# Notice the provided "yield" keyword, which allows calling the
# node that is being wrapped.
}
}

