% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageRunnerNode-run.R
\name{stageRunnerNode_run}
\alias{stageRunnerNode_run}
\title{Execute the callable of a stageRunnerNode.}
\usage{
stageRunnerNode_run(..., .cached_env = NULL, .callable = self$callable)
}
\arguments{
\item{...}{additional arguments to the \code{callable}. This allows,
stagerunner stages to be uniformly parametrized (for example,
if all stages should have a \code{verbose} parameter.}

\item{.cached_env}{An internal helper that passes the cached environment
to be used for storing the results of this execution.}

\item{.callable}{Another internal helper used for some recursive
metaprogramming.}
}
\value{
\code{TRUE} if the execution was successful, or an error otherwise.
}
\description{
Execute the callable of a stageRunnerNode.
}
