% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stageRunnerNode-transform.R
\name{stageRunnerNode_transform}
\alias{stageRunnerNode_transform}
\title{Transform a stageRunnerNode according to a functional.}
\usage{
stageRunnerNode_transform(transformation)
}
\arguments{
\item{transformation}{function. An arity-1 function which takes the
\code{callable} of a \code{stageRunnerNode} and transforms it
into another callable (i.e. a function or a stagerunner). If the
original \code{callable} is a stagerunner, its terminal nodes in
turn will be transformed recursively.}
}
\value{
The transformed callable.
}
\description{
Transform a stageRunnerNode according to a functional.
}
\examples{
\dontrun{
increment <- 1
adder     <- function(x) x + increment
node      <- stageRunnerNode$new(function(e) print(adder(1)))
node$transform(function(fn) {
  environment(fn)$increment <- environment(fn)$increment + 1; fn
})
node$run() # Prints 3, rather than 2
}
}

