% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/tree_skeleton.r
\name{treeSkeleton__.parent_index}
\alias{treeSkeleton__.parent_index}
\title{Find the index of the current object in the children of its parent.}
\usage{
treeSkeleton__.parent_index()
}
\description{
Find the index of the current object in the children of its parent.
}

