% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/tree_skeleton.r
\name{treeSkeleton__children}
\alias{treeSkeleton__children}
\title{Find the children of the current object wrapped in treeSkeletons.}
\usage{
treeSkeleton__children()
}
\description{
Find the children of the current object wrapped in treeSkeletons.
}

