% Generated by roxygen2 (4.0.1): do not edit by hand
\name{treeSkeleton__initialize}
\alias{treeSkeleton__initialize}
\title{Initialize a treeSkeleton object.}
\usage{
treeSkeleton__initialize(object, parent_caller = "parent",
  children_caller = "children")
}
\arguments{
\item{object}{ANY. If a reference class object, then \code{parent_caller}
and \code{children_caller} will refer to reference class methods.
If an attribute on the object with names of \code{children_caller} and
\code{parent_caller} exists, those will be used. Otherwise, the
generic methods will be used.}

\item{parent_caller}{character. The name of the reference class method
that returns the parent object, if the object was a node in a tree
structure.}

\item{children_caller}{character. The name of the reference class method
that returns the child objects, if the object was a node in a tree
structure.}
}
\value{
a treeSkeleton object.
}
\description{
treeSkeleton objects allow you to traverse a reference class object
as if it had a tree structure, merely by knowing how to call parent
or child nodes.
}

