% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/tree_skeleton.r
\name{treeSkeleton__parent}
\alias{treeSkeleton__parent}
\title{Find the parent of the current object wrapped in a treeSkeleton.}
\usage{
treeSkeleton__parent()
}
\description{
Find the parent of the current object wrapped in a treeSkeleton.
}

