% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treeSkeleton-predecessor.R
\name{treeSkeleton__predecessor}
\alias{treeSkeleton__predecessor}
\title{Attempt to find the predecessor of the current node.}
\usage{
treeSkeleton__predecessor(index = NULL)
}
\arguments{
\item{index}{integer. If specified, this is the index of the current node
in the children of its parent. (Sometimes, this cannot be computed
automatically, and should usually be provided.)}
}
\value{
predecessor for the wrapped object.
}
\description{
Attempt to find the predecessor of the current node.
}

