% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treeSkeleton.R
\name{treeSkeleton__root}
\alias{treeSkeleton__root}
\title{Find the root node of the tree (the only one with no parent).}
\usage{
treeSkeleton__root()
}
\value{
The root node of the tree or NULL if empty tree.
}
\description{
Find the root node of the tree (the only one with no parent).
}

