% Generated by roxygen2 (4.0.0): do not edit by hand
\name{treeSkeleton__successor}
\alias{treeSkeleton__successor}
\title{Attempt to find the successor of the current node.}
\arguments{
\item{index}{integer. If specified, this is the index of the current node
in the children of its parent. (Sometimes, this cannot be computed
automatically, and should usually be provided.)}
}
\value{
successor for the wrapped object.
}
\description{
Attempt to find the successor of the current node.
}

