% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boolean_fill}
\alias{boolean_fill}
\title{Fill a nested logical list with TRUEs before or after the first TRUE}
\usage{
boolean_fill(el, forward = TRUE)
}
\arguments{
\item{el}{list. A nested list of logicals with exactly one entry TRUE.}

\item{forward}{logical. FALSE for backwards, and TRUE for forwards.
The default is TRUE.}
}
\value{
the filled list
}
\description{
This is a helper function to implement the \code{to} parameter
in the \code{run} method on a stageRunner object.
}

