% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boolean_fill.r
\name{boolean_fill}
\alias{boolean_fill}
\title{Fill a nested logical list with TRUEs before or after the first TRUE}
\usage{
boolean_fill(el, forward = TRUE)
}
\arguments{
\item{el}{list. A nested list of logicals with exactly one entry \code{TRUE}.}

\item{forward}{logical. \code{FALSE} for backwards, and \code{TRUE} for forwards.
The default is \code{TRUE}.}
}
\value{
the filled list
}
\description{
This is a helper function to implement the \code{to} parameter
in the \code{run} method on a stageRunner object.
}

