% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_env.R
\name{copy_env}
\alias{copy_env}
\title{Copy one environment into another recursively.}
\usage{
copy_env(to, from)
}
\arguments{
\item{to}{environment. The new environment.}

\item{from}{environment. The old environment.}
}
\description{
Copy one environment into another recursively.
}
\note{
Both \code{to} and \code{from} must be pre-existing environments
  or this function will error.
}
