% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner_R6.R
\name{is.stagerunner}
\alias{is.stagerunner}
\title{Check whether an R object is a stageRunner object}
\usage{
is.stagerunner(obj)
}
\arguments{
\item{obj}{any object.}
}
\value{
\code{TRUE} if the object is of class
   \code{stageRunner}, \code{FALSE} otherwise.
}
\description{
Check whether an R object is a stageRunner object
}

