% Generated by roxygen2 (4.0.0): do not edit by hand
\name{show_message}
\alias{show_message}
\title{Show a progress message.}
\usage{
show_message(stage_names, stage_index, begin = TRUE, nested = FALSE,
  depth = 1)
}
\arguments{
\item{stage_names}{character.}

\item{stage_index}{integer.}

\item{begin}{logical. Whether we are showing the begin or end message.}

\item{nested}{logical. Whether or not this is a nested stage (i.e.
contains another stageRunner).}

\item{depth}{integer. How many tabs to space by (for nested stages).}
}
\value{
the message to standard output
}
\description{
Show a progress message.
}
\examples{
\dontrun{
show_message(c('one', 'two'), 2) # Will print "Beginning one stage..."
}
}

