% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stage_runner-package.R
\docType{package}
\name{stagerunner}
\alias{stagerunner}
\alias{stagerunner-package}
\title{stagerunner: in-memory reproducible data preparation and modeling}
\description{
stagerunner is an attempt to define a notion of data munging that includes
\emph{history}. By writing your code as a stagerunner instead of a
collection of functions, three key advantages should become clear:
}
\details{
\itemize{
\item Clarity will emerge in code that is intended to execute a sequence
  of operations that aims to produce a final result.
\item Reproducibility of interactive munging steps is possible without
  re-executing your analysis from scratch.
\item Modularity and extensibility becomes free of charge: methods like
  \code{around} and \code{transform} allow you to apply the same operation
  to your entire modeling procedure, simplifying progress monitoring and
  debugging.
}

Although originally intended for clarifying the modeling process,
stagerunners have much more general applicability. To learn more,
begin with the vignettes: \code{browseVignettes(package = "stagerunner")}.
}
\author{
Robert Krzyzanowski <\url{http://syberia.io}>
}
\references{
Full documentation and demos: \url{http://robertzk.github.io/stagerunner/};
  FAQ's: \url{http://robertzk.github.io/stagerunner/faq/}
}
\seealso{
The core function in this package: \code{\link{stagerunner}}. It
  defines the constructor creating stagerunner objects that allow you to
  wrap a complicated modeling procedure into an organized hierarchy.
}

