% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/stage_runner.r
\name{stageRunner}
\alias{stageRunner}
\title{Stage runner is a reference class for parametrizing and executing
a linear sequence of actions.}
\description{
Stage runner is a reference class for parametrizing and executing
a linear sequence of actions.
}

