% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stageRunner}
\alias{stageRunner}
\title{Stage runner is a reference class for parametrizing and executing
a linear sequence of actions.}
\description{
Stage runner is a reference class for parametrizing and executing
a linear sequence of actions.
}

