% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagerunner.R
\docType{class}
\name{stageRunner}
\alias{stageRunner}
\alias{stageRunner_}
\alias{stageRunner}
\alias{stagerunner}
\title{Stagerunners are parametrized sequences of linear execution.}
\usage{
stageRunner_

stageRunner(...)

stagerunner(...)
}
\description{
Stagerunners are parametrized sequences of linear execution.
}
\keyword{datasets}
