% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner.R
\docType{class}
\name{stageRunner}
\alias{stageRunner}
\alias{stageRunner_}
\alias{stagerunner}
\title{Stagerunners are parametrized sequences of linear execution.}
\format{\preformatted{function (...)  
 - attr(*, "srcref")=Class 'srcref'  atomic [1:8] 80 3 80 41 3 41 80 80
  .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x107680ab0> 
 - attr(*, "class")= chr "stageRunner_"
}}
\usage{
stageRunner_

stageRunner

stagerunner
}
\description{
Stagerunners are parametrized sequences of linear execution.
}
\keyword{datasets}

