% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stage_runner-package.R
\docType{package}
\name{stagerunner}
\alias{stagerunner}
\alias{stagerunner-package}
\title{StageRunner is a package for parametrizing linear sequences of actions.}
\description{
StageRunner is a package for parametrizing linear sequences of actions.
}

