<<<<<<< Updated upstream
\docType{class}
=======
% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{function}
>>>>>>> Stashed changes
\name{stageRunner}
\alias{stageRunner}
\title{Stage runner is a reference class for parametrizing and executing
a linear sequence of actions.}
\description{
Stage runner is a reference class for parametrizing and
executing a linear sequence of actions.
}

