% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stage_runner.r
\name{stageRunner__.set_parents}
\alias{stageRunner__.set_parents}
\title{Set all parents for this stageRunner, and recursively}
\usage{
stageRunner__.set_parents()
}
\description{
Set all parents for this stageRunner, and recursively
}

