% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/objectdiff.R
\name{stageRunner__.set_prefixes}
\alias{stageRunner__.set_prefixes}
\title{Set all prefixes for child stageRunners.}
\usage{
stageRunner__.set_prefixes(prefix = "")
}
\arguments{
\item{prefix}{character. The prefix to assign to this stageRunner.}
}
\description{
When a stageRunner is used in conjunction with an
\code{objectdiff::tracked_environment}, we need to remember
the full nested tree structure. This function sets up the
\code{prefix} member of each sub-stageRunner recursively to enable
correct remembering functionality.
}

