% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/stage_runner.r
\name{stageRunner__coalesce}
\alias{stageRunner__coalesce}
\title{Coalescing a stageRunner object is taking another stageRunner object
with similar stage names and replacing the latter's cached environments
with the former's.}
\usage{
stageRunner__coalesce(other_runner)
}
\arguments{
\item{other_runner}{stageRunner. Another stageRunner from which to coalesce.}
}
\description{
Coalescing a stageRunner object is taking another stageRunner object
with similar stage names and replacing the latter's cached environments
with the former's.
}

