% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/stage_runner.r
\name{stageRunner__show}
\alias{stageRunner__show}
\title{Generic for printing stageRunner objects.}
\usage{
stageRunner__show(indent = 0)
}
\arguments{
\item{indent}{integer. Internal parameter for keeping track of nested
indentation level.}
}
\description{
Generic for printing stageRunner objects.
}

