% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/stage_runner.r
\name{stageRunner__transform}
\alias{stageRunner__transform}
\title{Transform the callable's of the terminal nodes of a stageRunner.}
\usage{
stageRunner__transform(transformation)
}
\arguments{
\item{transformation}{function. The function which transforms one callable
into another.}
}
\description{
Every terminal node in a stageRunner is of type stageRunnerNode.
These each have a callable, and this method transforms those
callables in the way given by the first argument.
}

