% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stagerunner-current_stage.R
\name{stageRunner_current_stage}
\alias{stageRunner_current_stage}
\title{}
\usage{
stageRunner_current_stage()
}
\value{
a character stage key giving the latest executed stage.
  If the stageRunner does not have caching enabled, this will
  always return the first stage key (`'1'`).
}
\description{

}

