% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stage_runner.R, R/stagerunner_R6.R
\docType{class}
\name{stageRunnerNode}
\alias{.stageRunnerNode}
\alias{stageRunnerNode}
\title{Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "stageRunnerNode"
 $ public    :List of 15
  ..$ callable    : NULL
  ..$ cached_env  : NULL
  ..$ .context    : NULL
  ..$ .parent     : NULL
  ..$ executed    : logi FALSE
  ..$ initialize  :function (.callable, .context = NULL)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 724 18 728 5 18 5 724 728
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1060c6510> 
  ..$ run         :function (..., .cached_env = NULL, .callable = callable)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 729 11 743 5 11 5 729 743
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1060c6510> 
  ..$ around      :function (other_node)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 746 14 774 5 14 5 746 774
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1060c6510> 
  ..$ overlay     :function (other_node, label = NULL, flat = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 776 15 792 5 15 5 776 792
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1060c6510> 
  ..$ transform   :function (transformation)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 793 17 796 5 17 5 793 796
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1060c6510> 
  ..$ was_executed:function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 797 20 797 42 20 42 797 797
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1060c6510> 
  ..$ parent      :function (`*VALUE*` = NULL)  
  .. ..- attr(*, "srcref")= language 3L(5L, 5L, 44L, 5L, 44L, 3L, 5L)
  ..$ children    :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 799 16 799 32 16 32 799 799
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1060c6510> 
  ..$ show        :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 800 16 800 87 16 87 800 800
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1060c6510> 
  ..$ index       :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 803 16 807 5 16 5 803 807
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1060c6510> 
 $ private   : NULL
 $ active    : NULL
 $ inherit   : NULL
 $ parent_env:<environment: namespace:stagerunner> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
stageRunnerNode
}
\arguments{
\item{fn}{function. This will be wrapped in an environment.}

\item{parent_obj}{stageRunner. The enclosing stageRunner object.}

\item{parent_env}{environment. The parent environment of the created
\code{stageRunnerNode} object. The default is the calling
environment (i.e., \code{parent.frame()}).}

\item{fn}{function. This will be wrapped in an environment.}

\item{parent_obj}{stageRunner. The enclosing stageRunner object.}

\item{parent_env}{environment. The parent environment of the created
\code{stageRunnerNode} object. The default is the calling
environment (i.e., \code{parent.frame()}).}
}
\value{
an environment with some additional attributes for
  navigating in a tree-like structure.

an environment with some additional attributes for
  navigating in a tree-like structure.
}
\description{
Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).

Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).
}
\keyword{datasets}

