% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner_R6.R
\docType{class}
\name{stageRunnerNode}
\alias{stageRunnerNode}
\alias{stageRunnerNode_}
\title{Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "stageRunnerNode"
 $ public    :List of 15
  ..$ callable    : NULL
  ..$ cached_env  : NULL
  ..$ .context    : NULL
  ..$ .parent     : NULL
  ..$ executed    : logi FALSE
  ..$ initialize  :function (.callable, .context = NULL)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 733 18 736 5 18 5 733 736
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
  ..$ run         :function (..., .cached_env = NULL, .callable = self$callable)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 737 11 751 5 11 5 737 751
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
  ..$ around      :function (other_node)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 754 14 782 5 14 5 754 782
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
  ..$ overlay     :function (other_node, label = NULL, flat = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 784 15 800 5 15 5 784 800
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
  ..$ transform   :function (transformation)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 801 17 804 5 17 5 801 804
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
  ..$ was_executed:function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 805 20 805 42 20 42 805 805
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
  ..$ parent      :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 806 16 806 50 16 50 806 806
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
  ..$ children    :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 807 16 807 32 16 32 807 807
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
  ..$ show        :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 808 16 808 87 16 87 808 808
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
  ..$ index       :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 811 16 815 5 16 5 811 815
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10dd188e8> 
 $ private   : NULL
 $ active    : NULL
 $ inherit   : NULL
 $ parent_env:<environment: namespace:stagerunner> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
stageRunnerNode_
}
\arguments{
\item{fn}{function. This will be wrapped in an environment.}

\item{parent_obj}{stageRunner. The enclosing stageRunner object.}

\item{parent_env}{environment. The parent environment of the created
\code{stageRunnerNode} object. The default is the calling
environment (i.e., \code{parent.frame()}).}
}
\value{
an environment with some additional attributes for
  navigating in a tree-like structure.
}
\description{
Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).
}
\keyword{datasets}

