% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageRunnerNode.R
\docType{class}
\name{stageRunnerNode}
\alias{is.stageRunnerNode}
\alias{stageRunnerNode}
\alias{stageRunnerNode_}
\title{Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).}
\usage{
stageRunnerNode_

is.stageRunnerNode(obj)
}
\arguments{
\item{obj}{ANY. An object to test for class \code{stageRunnerNode}.}
}
\description{
Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).
}
\keyword{datasets}

