% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stagerunner_R6.R
\docType{class}
\name{stageRunnerNode}
\alias{stageRunnerNode}
\alias{stageRunnerNode_}
\title{Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).}
\format{\preformatted{List of 8
 $ new       :function (...)  
 $ classname : chr "stageRunnerNode"
 $ public    :List of 15
  ..$ callable    : NULL
  ..$ .cached_env : NULL
  ..$ .context    : NULL
  ..$ .parent     : NULL
  ..$ executed    : logi FALSE
  ..$ initialize  :function (.callable, .context = NULL)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 727 18 732 5 18 5 727 732
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
  ..$ run         :function (..., .cached_env = NULL, .callable = self$callable)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 734 11 748 5 11 5 734 748
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
  ..$ around      :function (other_node)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 751 14 779 5 14 5 751 779
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
  ..$ overlay     :function (other_node, label = NULL, flat = FALSE)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 781 15 797 5 15 5 781 797
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
  ..$ transform   :function (transformation)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 798 17 801 5 17 5 798 801
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
  ..$ was_executed:function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 802 20 802 47 20 47 802 802
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
  ..$ parent      :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 803 16 803 50 16 50 803 803
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
  ..$ children    :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 804 16 804 32 16 32 804 804
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
  ..$ show        :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 805 16 805 92 16 92 805 805
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
  ..$ index       :function ()  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 808 16 812 5 16 5 808 812
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x10f79c958> 
 $ private   : NULL
 $ active    : NULL
 $ inherit   : NULL
 $ parent_env:<environment: namespace:stagerunner> 
 $ lock      : logi TRUE
 - attr(*, "class")= chr "R6ClassGenerator"
}}
\usage{
stageRunnerNode_
}
\arguments{
\item{fn}{function. This will be wrapped in an environment.}

\item{parent_obj}{stageRunner. The enclosing stageRunner object.}

\item{parent_env}{environment. The parent environment of the created
\code{stageRunnerNode} object. The default is the calling
environment (i.e., \code{parent.frame()}).}
}
\value{
an environment with some additional attributes for
  navigating in a tree-like structure.
}
\description{
Stagerunner nodes are environment wrappers around individual stages
(i.e. functions) in order to track meta-data (e.g., for caching).
}
\keyword{datasets}

