% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/treeSkeleton.R
\name{treeSkeleton__children}
\alias{treeSkeleton__children}
\title{Find the children of the current object wrapped in treeSkeletons.}
\usage{
treeSkeleton__children()
}
\description{
Find the children of the current object wrapped in treeSkeletons.
}

