% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSkeleton.R
\name{treeSkeleton__first_leaf}
\alias{treeSkeleton__first_leaf}
\title{Find the first leaf in a tree.}
\usage{
treeSkeleton__first_leaf()
}
\value{
The first leaf, that is, the first terminal child node.
}
\description{
Find the first leaf in a tree.
}

