% Generated by roxygen2 (4.0.0): do not edit by hand
\name{treeSkeleton__root}
\alias{treeSkeleton__root}
\title{Find the root node of the tree (the only one with no parent).}
\value{
The root node of the tree or NULL if empty tree.
}
\description{
Find the root node of the tree (the only one with no parent).
}

